move env.ts to within src/
